requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
django-filter>=23.0
requests>=2.31
python-decouple>=3.8
Pillow>=10.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
celery>=5.3
redis>=5.0
gunicorn>=21.0
django-redis>=5.3
django-ratelimit>=4.1
